Add files via upload
problem statement:he ‘Hair Salon.csv’ dataset contains various variables used for the
context of Market Segmentation. This case study is based on various parameters of a salon
chain of hair products. You are expected to do a Principal Component Analysis for this case
study according to the instructions given in the rubric. Kindly refer to
the PCA_Data_Dictionary.jpg file for the Data Dictionary of the Dataset.
Note: This particular dataset contains the target variable satisfaction as well. Please drop
this variable before doing a Principal Component Analysis